+++ internal/controller/alb/forwardingrule/forwardingrule.go
@@ -238,25 +238,25 @@ func (c *externalForwardingRule) Update(ctx context.Context, mg resource.Managed
 	return managed.ExternalUpdate{}, nil
 }
 
-func (c *externalForwardingRule) Delete(ctx context.Context, mg resource.Managed) error {
+func (c *externalForwardingRule) Delete(ctx context.Context, mg resource.Managed) (managed.ExternalDelete, error) {
 	cr, ok := mg.(*v1alpha1.ForwardingRule)
 	if !ok {
-		return errors.New(errNotForwardingRule)
+		return managed.ExternalDelete{}, errors.New(errNotForwardingRule)
 	}
 	cr.SetConditions(xpv1.Deleting())
 	if cr.Status.AtProvider.State == string(ionoscloud.STATE_DESTROYING) || cr.Status.AtProvider.State == string(ionoscloud.STATE_BUSY) {
-		return nil
+		return managed.ExternalDelete{}, nil
 	}
 	apiResponse, err := c.service.DeleteForwardingRule(ctx, cr.Spec.ForProvider.DatacenterCfg.DatacenterID,
 		cr.Spec.ForProvider.ALBCfg.ApplicationLoadBalancerID, cr.Status.AtProvider.ForwardingRuleID)
 	if err != nil {
 		retErr := fmt.Errorf("failed to delete application load balancer forwarding rule. error: %w", err)
-		return compute.ErrorUnlessNotFound(apiResponse, retErr)
+		return managed.ExternalDelete{}, compute.ErrorUnlessNotFound(apiResponse, retErr)
 	}
 	if err = compute.WaitForRequest(ctx, c.service.GetAPIClient(), apiResponse); err != nil {
-		return err
+		return managed.ExternalDelete{}, err
 	}
-	return nil
+	return managed.ExternalDelete{}, nil
 }
 
 func (c *externalForwardingRule) getIPSet(ctx context.Context, cr *v1alpha1.ForwardingRule) (string, error) {
@@ -274,4 +274,9 @@ func (c *externalForwardingRule) getIPSet(ctx context.Context, cr *v1alpha1.Forw
 		return ipsCfg[0], nil
 	}
 	return "", nil
+}
+
+// Disconnect does nothing because there are no resources to release. Needs to be implemented starting from crossplane-runtime v0.17
+func (c *externalForwardingRule) Disconnect(_ context.Context) error {
+	return nil
 }