+++ internal/controller/compute/datacenter/datacenter.go
@@ -211,24 +211,29 @@ func (c *externalDatacenter) Update(ctx context.Context, mg resource.Managed) (m
 	return managed.ExternalUpdate{}, nil
 }
 
-func (c *externalDatacenter) Delete(ctx context.Context, mg resource.Managed) error {
+func (c *externalDatacenter) Delete(ctx context.Context, mg resource.Managed) (managed.ExternalDelete, error) {
 	cr, ok := mg.(*v1alpha1.Datacenter)
 	if !ok {
-		return errors.New(errNotDatacenter)
+		return managed.ExternalDelete{}, errors.New(errNotDatacenter)
 	}
 
 	cr.SetConditions(xpv1.Deleting())
 	if cr.Status.AtProvider.State == compute.DESTROYING {
-		return nil
+		return managed.ExternalDelete{}, nil
 	}
 
 	apiResponse, err := c.service.DeleteDatacenter(ctx, cr.Status.AtProvider.DatacenterID)
 	if err != nil {
 		retErr := fmt.Errorf("failed to delete datacenter. error: %w", err)
-		return compute.ErrorUnlessNotFound(apiResponse, retErr)
+		return managed.ExternalDelete{}, compute.ErrorUnlessNotFound(apiResponse, retErr)
 	}
 	if err = compute.WaitForRequest(ctx, c.service.GetAPIClient(), apiResponse); err != nil {
-		return err
+		return managed.ExternalDelete{}, err
 	}
+	return managed.ExternalDelete{}, nil
+}
+
+// Disconnect does nothing because there are no resources to release. Needs to be implemented starting from crossplane-runtime v0.17
+func (c *externalDatacenter) Disconnect(_ context.Context) error {
 	return nil
 }