+++ internal/controller/compute/ipblock/ipblock.go
@@ -215,24 +215,29 @@ func (c *externalIPBlock) Update(ctx context.Context, mg resource.Managed) (mana
 	return managed.ExternalUpdate{}, nil
 }
 
-func (c *externalIPBlock) Delete(ctx context.Context, mg resource.Managed) error {
+func (c *externalIPBlock) Delete(ctx context.Context, mg resource.Managed) (managed.ExternalDelete, error) {
 	cr, ok := mg.(*v1alpha1.IPBlock)
 	if !ok {
-		return errors.New(errNotIPBlock)
+		return managed.ExternalDelete{}, errors.New(errNotIPBlock)
 	}
 
 	cr.SetConditions(xpv1.Deleting())
 	if cr.Status.AtProvider.State == compute.DESTROYING {
-		return nil
+		return managed.ExternalDelete{}, nil
 	}
 
 	apiResponse, err := c.service.DeleteIPBlock(ctx, cr.Status.AtProvider.IPBlockID)
 	if err != nil {
 		retErr := fmt.Errorf("failed to delete ipBlock. error: %w", err)
-		return compute.ErrorUnlessNotFound(apiResponse, retErr)
+		return managed.ExternalDelete{}, compute.ErrorUnlessNotFound(apiResponse, retErr)
 	}
 	if err = compute.WaitForRequest(ctx, c.service.GetAPIClient(), apiResponse); err != nil {
-		return err
+		return managed.ExternalDelete{}, err
 	}
+	return managed.ExternalDelete{}, nil
+}
+
+// Disconnect does nothing because there are no resources to release. Needs to be implemented starting from crossplane-runtime v0.17
+func (c *externalIPBlock) Disconnect(_ context.Context) error {
 	return nil
 }