+++ docs/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## [1.1.4]
 - **Fixes**:
  - Volume should populate the status ID from the bootvolume of the server, because it cannot store it on create. This can cause larger wait times for volume attach.
-
+ - Sss - re-create datavolumes in case of error.
+ - Sss - copy to configmap, don't just overwrite(edge case in case crossplane provider reboots)
+ - Sss add additional logs for debugging
+ - Fix `volume` yaml example
+ - docs update
 ## [1.1.3]
 - **Features**:
     - Enable `publishConnectionDetails` option for s3 key, compute user and k8s clusters